requirements.txt
google-generativeai>=0.3.0
pdf2image==1.16.3
pillow==10.3.0
pypdf==3.17.0
python-dotenv==1.0.0
